+++ backend/app/tasks/_frame_deployer.py
@@ -1,10 +1,14 @@
 
 from datetime import datetime
+import hashlib
 import json
 import math
+import os
 from pathlib import Path
 import random
+import re
 import shlex
+import shutil
 import string
 import tempfile
 from typing import Optional
@@ -16,6 +20,12 @@ from app.models.frame import Frame, get_frame_json
 from app.models.log import new_log as log
 from app.utils.local_exec import exec_local_command
 from app.utils.remote_exec import upload_file, run_command
+from app.drivers.waveshare import write_waveshare_driver_nim, get_variant_folder
+from app.drivers.devices import drivers_for_frame
+from app.models import get_apps_from_scenes
+from app.codegen.drivers_nim import write_drivers_nim
+from app.codegen.scene_nim import write_scene_nim, write_scenes_nim
+from app.tasks.utils import find_nimbase_file
 
 BYTES_PER_MB   = 1_048_576
 DEFAULT_CHUNK  = 25 * BYTES_PER_MB
@@ -191,6 +201,12 @@ class FrameDeployer:
 
         return len(missing)
 
+    async def get_hostname(self) -> str:
+        hostname_out: list[str] = []
+        await self.exec_command("hostname", hostname_out)  # remote hostname for flake attribute
+        target_host = hostname_out[0].strip() or "frame-unknown"
+        return target_host
+
     async def get_distro(self) -> str:
         distro_out: list[str] = []
         await self.exec_command(
@@ -240,3 +256,217 @@ class FrameDeployer:
 
     async def disable_service(self, service_name: str) -> None:
         await self.exec_command(f"sudo systemctl disable {service_name}.service", raise_on_error=False)
+
+    async def make_local_modifications(self, source_dir: str):
+        frame = self.frame
+        shutil.rmtree(os.path.join(source_dir, "src", "scenes"), ignore_errors=True)
+        os.makedirs(os.path.join(source_dir, "src", "scenes"), exist_ok=True)
+
+        for node_id, sources in get_apps_from_scenes(list(frame.scenes)).items():
+            app_id = "nodeapp_" + node_id.replace('-', '_')
+            app_dir = os.path.join(source_dir, "src", "apps", app_id)
+            os.makedirs(app_dir, exist_ok=True)
+            for filename, code in sources.items():
+                with open(os.path.join(app_dir, filename), "w") as f:
+                    f.write(code)
+
+        for scene in frame.scenes:
+            try:
+                scene_source = write_scene_nim(frame, scene)
+                safe_id = re.sub(r'\W+', '', scene.get('id', 'default'))
+                with open(os.path.join(source_dir, "src", "scenes", f"scene_{safe_id}.nim"), "w") as f:
+                    f.write(scene_source)
+            except Exception as e:
+                await self.log("stderr",
+                        f"Error writing scene \"{scene.get('name','')}\" "
+                        f"({scene.get('id','default')}): {e}")
+                raise
+
+        with open(os.path.join(source_dir, "src", "scenes", "scenes.nim"), "w") as f:
+            source = write_scenes_nim(frame)
+            f.write(source)
+            if frame.debug:
+                await self.log("stdout", f"Generated scenes.nim (showing because debug=true):\n{source}")
+
+        drivers = drivers_for_frame(frame)
+        with open(os.path.join(source_dir, "src", "drivers", "drivers.nim"), "w") as f:
+            source = write_drivers_nim(drivers)
+            f.write(source)
+            if frame.debug:
+                await self.log("stdout", f"Generated drivers.nim:\n{source}")
+
+        if drivers.get("waveshare"):
+            with open(os.path.join(source_dir, "src", "drivers", "waveshare", "driver.nim"), "w") as wf:
+                source = write_waveshare_driver_nim(drivers)
+                wf.write(source)
+                if frame.debug:
+                    await self.log("stdout", f"Generated waveshare driver:\n{source}")
+
+
+    def create_local_source_folder(self, temp_dir: str) -> str:
+        source_dir = os.path.join(temp_dir, "frameos")
+        os.makedirs(source_dir, exist_ok=True)
+        shutil.copytree("../frameos", source_dir, dirs_exist_ok=True)
+        return source_dir
+
+
+    async def create_local_build_archive(
+        self,
+        build_dir: str,
+        source_dir: str,
+        arch: str
+    ):
+        db = self.db
+        redis = self.redis
+        frame = self.frame
+        build_id = self.build_id
+        nim_path = self.nim_path
+        temp_dir = self.temp_dir
+
+        drivers = drivers_for_frame(frame)
+        if inkyPython := drivers.get('inkyPython'):
+            vendor_folder = inkyPython.vendor_folder or ""
+            os.makedirs(os.path.join(build_dir, "vendor"), exist_ok=True)
+            shutil.copytree(
+                f"../frameos/vendor/{vendor_folder}/",
+                os.path.join(build_dir, "vendor", vendor_folder),
+                dirs_exist_ok=True
+            )
+            shutil.rmtree(os.path.join(build_dir, "vendor", vendor_folder, "env"), ignore_errors=True)
+            shutil.rmtree(os.path.join(build_dir, "vendor", vendor_folder, "__pycache__"), ignore_errors=True)
+
+        if inkyHyperPixel2r := drivers.get('inkyHyperPixel2r'):
+            vendor_folder = inkyHyperPixel2r.vendor_folder or ""
+            os.makedirs(os.path.join(build_dir, "vendor"), exist_ok=True)
+            shutil.copytree(
+                f"../frameos/vendor/{vendor_folder}/",
+                os.path.join(build_dir, "vendor", vendor_folder),
+                dirs_exist_ok=True
+            )
+            shutil.rmtree(os.path.join(build_dir, "vendor", vendor_folder, "env"), ignore_errors=True)
+            shutil.rmtree(os.path.join(build_dir, "vendor", vendor_folder, "__pycache__"), ignore_errors=True)
+
+        await self.log("stdout",
+                "- No cross compilation. Generating source code for compilation on frame.")
+
+        cpu = await self.arch_to_nim_cpu(arch)
+        debug_options = "--lineTrace:on" if frame.debug else ""
+        cmd = (
+            f"cd {source_dir} && nimble assets -y && nimble setup && "
+            f"{nim_path} compile --os:linux --cpu:{cpu} "
+            f"--compileOnly --genScript --nimcache:{build_dir} "
+            f"{debug_options} src/frameos.nim 2>&1"
+        )
+
+        status, out, err = await exec_local_command(db, redis, frame, cmd)
+        if status != 0:
+            lines = (out or "").split("\n")
+            filtered = [ln for ln in lines if ln.strip()]
+            if filtered:
+                last_line = filtered[-1]
+                match = re.match(r'^(.*\.nim)\((\d+), (\d+)\),?.*', last_line)
+                if match:
+                    fn = match.group(1)
+                    line_nr = int(match.group(2))
+                    column = int(match.group(3))
+                    source_abs = os.path.realpath(source_dir)
+                    final_path = os.path.realpath(os.path.join(source_dir, fn))
+                    if os.path.commonprefix([final_path, source_abs]) == source_abs:
+                        rel_fn = final_path[len(source_abs) + 1:]
+                        with open(final_path, "r") as of:
+                            all_lines = of.readlines()
+                        await self.log("stdout", f"Error in {rel_fn}:{line_nr}:{column}")
+                        await self.log("stdout", f"Line {line_nr}: {all_lines[line_nr - 1]}")
+                        await self.log("stdout", f".......{'.'*(column - 1 + len(str(line_nr)))}^")
+                    else:
+                        await self.log("stdout", f"Error in {fn}:{line_nr}:{column}")
+
+            raise Exception("Failed to generate frameos sources")
+
+        nimbase_path = find_nimbase_file(nim_path)
+        if not nimbase_path:
+            raise Exception("nimbase.h not found")
+
+        shutil.copy(nimbase_path, os.path.join(build_dir, "nimbase.h"))
+
+        if waveshare := drivers.get('waveshare'):
+            if waveshare.variant:
+                variant_folder = get_variant_folder(waveshare.variant)
+                util_files = ["Debug.h", "DEV_Config.c", "DEV_Config.h"]
+                for uf in util_files:
+                    shutil.copy(
+                        os.path.join(source_dir, "src", "drivers", "waveshare", variant_folder, uf),
+                        os.path.join(build_dir, uf)
+                    )
+
+                # color e-paper variants
+                if waveshare.variant in [
+                    "EPD_2in9b", "EPD_2in9c", "EPD_2in13b", "EPD_2in13c",
+                    "EPD_4in2b", "EPD_4in2c", "EPD_5in83b", "EPD_5in83c",
+                    "EPD_7in5b", "EPD_7in5c"
+                ]:
+                    c_file = re.sub(r'[bc]', 'bc', waveshare.variant)
+                    variant_files = [f"{waveshare.variant}.nim", f"{c_file}.c", f"{c_file}.h"]
+                else:
+                    variant_files = [f"{waveshare.variant}.nim", f"{waveshare.variant}.c", f"{waveshare.variant}.h"]
+
+                for vf in variant_files:
+                    shutil.copy(
+                        os.path.join(source_dir, "src", "drivers", "waveshare", variant_folder, vf),
+                        os.path.join(build_dir, vf)
+                    )
+
+        with open(os.path.join(build_dir, "Makefile"), "w") as mk:
+            script_path = os.path.join(build_dir, "compile_frameos.sh")
+            linker_flags = ["-pthread", "-lm", "-lrt", "-ldl"]
+            compiler_flags: list[str] = []
+            with open(script_path, "r") as sc:
+                lines_sc = sc.readlines()
+            for line in lines_sc:
+                if " -o frameos " in line and " -l" in line:
+                    linker_flags = [
+                        fl.strip() for fl in line.split(' ')
+                        if fl.startswith('-') and fl != '-o'
+                    ]
+                elif " -c " in line and not compiler_flags:
+                    compiler_flags = [
+                        fl for fl in line.split(' ')
+                        if fl.startswith('-') and not fl.startswith('-I')
+                        and fl not in ['-o', '-c', '-D']
+                    ]
+
+            with open(os.path.join(source_dir, "tools", "nimc.Makefile"), "r") as mf_in:
+                lines_make = mf_in.readlines()
+            for ln in lines_make:
+                if ln.startswith("LIBS = "):
+                    ln = "LIBS = -L. " + " ".join(linker_flags) + "\n"
+                if ln.startswith("CFLAGS = "):
+                    ln = "CFLAGS = " + " ".join([f for f in compiler_flags if f != '-c']) + "\n"
+                mk.write(ln)
+
+        archive_path = os.path.join(temp_dir, f"build_{build_id}.tar.gz")
+        zip_base = os.path.join(temp_dir, f"build_{build_id}")
+        shutil.make_archive(zip_base, 'gztar', temp_dir, f"build_{build_id}")
+        return archive_path
+
+    async def file_in_sync(
+        self: "FrameDeployer",
+        remote_path: str,
+        local_path: str,
+    ) -> bool:
+        """
+        Check if the remote path is the same as the local path.
+        """
+        def sha256(path: str) -> str:
+            h = hashlib.sha256()
+            with open(path, "rb") as fp:
+                for chunk in iter(lambda: fp.read(65536), b""):
+                    h.update(chunk)
+            return h.hexdigest()
+
+        remote_sha_out: list[str] = []
+        await self.exec_command(f"sha256sum {remote_path} 2>/dev/null | cut -d' ' -f1 || true", remote_sha_out)
+        remote_sha = (remote_sha_out[0].strip() if remote_sha_out else "")
+
+        local_sha = sha256(local_path)
+        return remote_sha == local_sha
